+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file. This projec
 - Album and track word clouds: Added more stop words
 - Further automations and improvements to release process
 - Artist graphs: Consistently trim long artist names
+- Play count distribution: Limit to 99th percentile of play counts
 
 ## [1.1.0] - 2025-07-19